requirements.txt
Flask==2.2.2
python-dotenv==0.21.0
Werkzeug==2.2.2
google-genai
gunicorn
PyPDF2
openai
